Ensure USER is set when starting services
When the flight service stack is started by
`flight-plugin-system-systemd-service`, the environment variable `USER`
is not set.  This can cause some of our serivces to fail to start.

Specifically, `flight-desktop-restapi` runs a `flight-desktop` version
check command on boot which fails if `USER` is not set.  If it fails,
`flight-desktop-restapi` fails to start.