Fix issue with stop incorrectly reporting failure
Prior to the `stop` command learning the `--force` flag we trusted the
exit status of the `stop.sh` script to be authoritative as to whether it
was successful.  With the introduction of the `--force` flag we need to
ignore the exit status of `stop.sh` and instead check if the process is
running or not.

This introduces a race condition if the process takes a short while to
terminate after the `stop.sh` script exits.  If the race condition is
lost, we report that the process did not get stopped.  But that may not
be the case.

This commit fixes the issue by waiting for a short period of time for
the process to terminate after the `stop.sh` script has exited.  Along
with fixing the issue detailed above, it also helps us to avoid being
overzealous in sending `TERM` when the `--force` flag is given.